verify_service/requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
python-jose==3.3.0
redis==5.2.0
pydantic-settings==2.5.2